[info] Split React components into proper modules
